+++ pkg/kubeapiserver/authenticator/config.go
@@ -30,15 +30,15 @@ import (
 	"k8s.io/apiserver/pkg/authentication/request/x509"
 	"k8s.io/apiserver/pkg/authentication/token/tokenfile"
 	"k8s.io/apiserver/pkg/authentication/user"
+	"k8s.io/apiserver/plugin/pkg/authenticator/password/keystone"
+	"k8s.io/apiserver/plugin/pkg/authenticator/password/passwordfile"
+	"k8s.io/apiserver/plugin/pkg/authenticator/request/basicauth"
+	"k8s.io/apiserver/plugin/pkg/authenticator/token/anytoken"
+	"k8s.io/apiserver/plugin/pkg/authenticator/token/oidc"
+	"k8s.io/apiserver/plugin/pkg/authenticator/token/webhook"
 	certutil "k8s.io/client-go/util/cert"
 	genericauthenticator "k8s.io/kubernetes/pkg/genericapiserver/authenticator"
 	"k8s.io/kubernetes/pkg/serviceaccount"
-	"k8s.io/kubernetes/plugin/pkg/auth/authenticator/password/keystone"
-	"k8s.io/kubernetes/plugin/pkg/auth/authenticator/password/passwordfile"
-	"k8s.io/kubernetes/plugin/pkg/auth/authenticator/request/basicauth"
-	"k8s.io/kubernetes/plugin/pkg/auth/authenticator/token/anytoken"
-	"k8s.io/kubernetes/plugin/pkg/auth/authenticator/token/oidc"
-	"k8s.io/kubernetes/plugin/pkg/auth/authenticator/token/webhook"
 
 	// Initialize all known client auth plugins.
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
