+++ staging/src/k8s.io/apiserver/plugin/pkg/authenticator/password/doc.go
@@ -14,7 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package keystone provides authentication via keystone.
-// For details about keystone and how to use the plugin, refer to
-// https://github.com/kubernetes/kubernetes.github.io/blob/master/docs/admin/authentication.md
-package keystone // import "k8s.io/kubernetes/plugin/pkg/auth/authenticator/password/keystone"
+// Package password contains authenticator.Password implementations
+package password // import "k8s.io/apiserver/plugin/pkg/authenticator/password"
