+++ staging/src/k8s.io/apiserver/plugin/pkg/authenticator/password/keystone/doc.go
@@ -14,5 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package authenticator contains implementations for pkg/auth/authenticator interfaces
-package authenticator // import "k8s.io/kubernetes/plugin/pkg/auth/authenticator"
+// Package keystone provides authentication via keystone.
+// For details about keystone and how to use the plugin, refer to
+// https://github.com/kubernetes/kubernetes.github.io/blob/master/docs/admin/authentication.md
+package keystone // import "k8s.io/apiserver/plugin/pkg/authenticator/password/keystone"
