+++ staging/src/k8s.io/apiserver/plugin/pkg/authenticator/password/keystone/keystone.go
@@ -25,6 +25,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/rackspace/gophercloud"
 	"github.com/rackspace/gophercloud/openstack"
+
 	netutil "k8s.io/apimachinery/pkg/util/net"
 	"k8s.io/apiserver/pkg/authentication/user"
 	certutil "k8s.io/client-go/util/cert"
