+++ staging/src/k8s.io/apiserver/plugin/pkg/authenticator/password/passwordfile/BUILD
