+++ staging/src/k8s.io/apiserver/plugin/pkg/authenticator/token/anytoken/BUILD
